Fixes Codebuild CI failures
